Add helper to `Sender`
