+++ common/src/test/java/org/astraea/common/metrics/collector/MetricsFetcherTest.java
@@ -23,15 +23,29 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutionException;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 import org.astraea.common.Utils;
+import org.astraea.common.admin.Admin;
+import org.astraea.common.consumer.Consumer;
+import org.astraea.common.consumer.Deserializer;
+import org.astraea.common.consumer.SeekStrategy;
 import org.astraea.common.metrics.BeanObject;
 import org.astraea.common.metrics.MBeanClient;
+import org.astraea.it.Service;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 public class MetricsFetcherTest {
+  static Service SERVICE = Service.builder().numberOfWorkers(0).build();
+
+  @AfterAll
+  static void close() {
+    SERVICE.close();
+  }
 
   @Test
   void testPublishAndClose() {
@@ -113,6 +127,34 @@ public class MetricsFetcherTest {
       Assertions.assertEquals(1, fetcher.clientIds().size());
       // the delay is too larger to see next update
       Mockito.verify(supplier, Mockito.times(1)).get();
+    }
+  }
+
+  @Test
+  void testTopic() throws InterruptedException, ExecutionException {
+    var testBean = new BeanObject("java.lang", Map.of("name", "n1"), Map.of("value", "v1"));
+    try (var topicSender = MetricsFetcher.Sender.topic(SERVICE.bootstrapServers())) {
+      topicSender.send(1, List.of(testBean));
+
+      // Test topic creation
+      try (var admin = Admin.of(SERVICE.bootstrapServers())) {
+        var topics = admin.topicNames(false).toCompletableFuture().get();
+        Assertions.assertEquals(1, topics.size());
+        Assertions.assertEquals("__1_broker_metrics", topics.stream().findAny().get());
+      }
+
+      // Test record sent
+      try (var consumer =
+          Consumer.forTopics(Set.of("__1_broker_metrics"))
+              .bootstrapServers(SERVICE.bootstrapServers())
+              .valueDeserializer(Deserializer.STRING)
+              .seek(SeekStrategy.DISTANCE_FROM_BEGINNING, 0)
+              .build()) {
+        var records =
+            consumer.poll(Duration.ofSeconds(5)).stream().collect(Collectors.toUnmodifiableList());
+        Assertions.assertEquals(1, records.size());
+        Assertions.assertEquals(testBean.toString(), records.get(0).value());
+      }
     }
   }
 }
